Test: extract write_fits function
This function was duplicated between many of the test scripts, with
slightly different versions present in testFilelistContainer and
in test_delayed_error_warning.  This commit extracts the function,
including the functionality from both of those test scripts, into a
new module.  Since this is now imported via a relative import it is no
longer possible to execute these tests as scripts -- they should be run
using the unittest test runner instead.