Update caom2repo_wrapper for direct XML access
The wrapper has been updated to take advantage of a version of
caom2repoClient which accepts and returns XML strings rather than
having to interact via files.  The process method now supplies a
CAOM-2 observation object rather than a filename.

The process method (context manager) needs to be able to provide the
existing observation, if it exists and also receive the updated
observation, or a newly created one if it doesn't exist (or if
fits2caom2 was used).  Therefore this commit also introduces a
box data type "ObservationWrapper" which is just used to contain
a CAOM-2 observation object.  There may be a more elegant solution
which would still allow the use of a context manager.