Attempt to remove excess parts from CAOM-2 records
fits2caom2 doesn't seem to remove parts left over from previous ingestions
which no longer correspond to FITS extensions which still exist.
Instead it just updates the parts corresponding to the extensions of the
FITS file it is given.  This commit therefore adds a step to remove these
excess parts before fits2caom2 is called.  (Doing it this way round allows
fits2caom2 to have the last word about the parts which should be present.)
In order to do this, a new instance variable "artifact_part_count" has
been added.  This is a dictionary which maps artifact URLs to numbers
of parts.  Whenever a FITS file is opened to have its header read,
the number of HDUs in the file is stored in this new dictionary.