Remove excess CAOM-2 record parts in num. order
The record parts (in the JCMT archive at least) are all numeric
(corresponding to the FITS extensions for reduced data), but are
not stored in numeric order.  Therefore we need to sort the list
of plane names (in numeric order) to be able to remove those
(with the highest number part names) which correspond to
FITS extensions which no longer exist.