Expose validation is_in_archive method
Renamed the (private) method _is_in_archive to is_in_archive (to
mark it as public) and altered it to raise an exception if the file
is missing from the archive.  This is because there isn't any code
using the delated_error_warning.in_archive method but there is code
in jcmt2caom2ingest which duplicates the simple check that a file is
in the archive.  This change will allow jcmt2caom2ingest to just call
is_in_archive instead.